+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## [Unreleased]
 
+## [2.6.0-alpha] - 2025-01-07
+
+### Added
+- Show indices for filenames.
+- Enter digits to jump to filename indices.
+
 ## [2.5.1-alpha] - 2025-01-05
 
 ### Added
@@ -89,7 +95,8 @@
   - **Edit**: Modify item text.
   - **Delete**: Delete an item.
 
-[Unreleased]: https://github.com/darthkeith/sieve-selector/compare/v2.5.1-alpha...HEAD
+[Unreleased]: https://github.com/darthkeith/sieve-selector/compare/v2.6.0-alpha...HEAD
+[2.6.0-alpha]: https://github.com/darthkeith/sieve-selector/compare/v2.5.1-alpha...v2.6.0-alpha
 [2.5.1-alpha]: https://github.com/darthkeith/sieve-selector/compare/v2.5.0-alpha...v2.5.1-alpha
 [2.5.0-alpha]: https://github.com/darthkeith/sieve-selector/compare/v2.4.0-alpha...v2.5.0-alpha
 [2.4.0-alpha]: https://github.com/darthkeith/sieve-selector/compare/v2.3.0-alpha...v2.4.0-alpha